+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Fixed extra space in Flex admin pages
     * Fixed folder creation with parent other than root [#66](https://github.com/trilbymedia/grav-plugin-flex-objects/issues/66)
     * Fixed task redirects in sub-folder multi-site environments
+    * Fixed typo in default permissions (should have been `admin.flex-objects`) [grav#2915](https://github.com/getgrav/grav/issues/2915)
 
 # v1.0.0-rc.15
 ## 07/22/2020