+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v1.0.0-rc.17
+## mm/dd/2020
+
+1. [](#bugfix)
+    * Fixed media uploads for objects which do not implement `FlexAuthorizeInterface`
+
 # v1.0.0-rc.16
 ## 09/01/2020